+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/utils/getRefreshTime.test.ts
@@ -5,7 +5,7 @@ const mockDate = new Date("2020-01-01");
 jest.useFakeTimers().setSystemTime(mockDate);
 describe("getRefreshTime", () => {
   it("returns default refresh time when no rates provided", () => {
-    expect(getRefreshTime(undefined)).toEqual(60000);
+    expect(getRefreshTime(undefined)).toEqual(20000);
   });
 
   it("returns the default refresh time when no rates have an expirationTime", () => {
@@ -15,7 +15,7 @@ describe("getRefreshTime", () => {
           rateId: "1",
         } as ExchangeRate,
       ]),
-    ).toEqual(60000);
+    ).toEqual(20000);
   });
 
   it("returns the a refresh time that is the earliest expirationTime in the list of rates if it's under 60s", () => {
@@ -24,17 +24,17 @@ describe("getRefreshTime", () => {
       getRefreshTime([
         {
           rateId: "1",
-          expirationTime: mockTimeSinceEpoch + 59000,
+          expirationTime: mockTimeSinceEpoch + 19000,
         },
         {
           rateId: "2",
-          expirationTime: mockTimeSinceEpoch + 58000,
+          expirationTime: mockTimeSinceEpoch + 18000,
         },
         {
           rateId: "3",
         },
       ] as ExchangeRate[]),
-    ).toEqual(58000);
+    ).toEqual(18000);
   });
 
   it("returns 60s when the earliest expirationTime in the list of rates is over 60s", () => {
@@ -43,16 +43,16 @@ describe("getRefreshTime", () => {
       getRefreshTime([
         {
           rateId: "1",
-          expirationTime: mockTimeSinceEpoch + 61000,
+          expirationTime: mockTimeSinceEpoch + 21000,
         },
         {
           rateId: "2",
-          expirationTime: mockTimeSinceEpoch + 62000,
+          expirationTime: mockTimeSinceEpoch + 22000,
         },
         {
           rateId: "3",
         },
       ] as ExchangeRate[]),
-    ).toEqual(60000);
+    ).toEqual(20000);
   });
 });