+++ apps/ledger-live-desktop/src/renderer/screens/manager/FirmwareUpdate/index.tsx
@@ -19,11 +19,12 @@ import { FakeLink } from "~/renderer/components/TopBanner";
 import { context } from "~/renderer/drawers/Provider";
 import { track } from "~/renderer/analytics/segment";
 import { LocalTracer } from "@ledgerhq/logs";
+import { useLocalizedUrl } from "~/renderer/hooks/useLocalizedUrls";
 
 type Props = {
   deviceInfo: DeviceInfo;
   device: Device;
-  setFirmwareUpdateOpened: (a: boolean) => void;
+  setPreventResetOnDeviceChange: (value: boolean) => void;
   disableFirmwareUpdate?: boolean;
   installed?: InstalledItem[];
   onReset: (a: string[]) => void;
@@ -56,7 +57,7 @@ const FirmwareUpdate = (props: Props) => {
   const {
     deviceInfo,
     device,
-    setFirmwareUpdateOpened,
+    setPreventResetOnDeviceChange,
     disableFirmwareUpdate,
     installed,
     firmware,
@@ -71,7 +72,7 @@ const FirmwareUpdate = (props: Props) => {
   const deviceSpecs = getDeviceModel(device.modelId);
   const isDeprecated = manager.firmwareUnsupported(device.modelId, deviceInfo);
   const [tracer] = useState(() => new LocalTracer("manager", { component: "FirmwareUpdate" }));
-
+  const contactSupportUrl = useLocalizedUrl(urls.contactSupport);
   const onDrawerClose = useCallback(() => {
     onReset((installed || []).map(({ name }) => name));
   }, [installed, onReset]);
@@ -81,11 +82,12 @@ const FirmwareUpdate = (props: Props) => {
   }, []);
 
   const onRequestClose = useCallback(() => {
+    setPreventResetOnDeviceChange(false);
     setDrawer();
     if (firmwareUpdateCompletedRef.current) {
       onReset([]);
     }
-  }, [onReset, setDrawer]);
+  }, [onReset, setDrawer, setPreventResetOnDeviceChange]);
 
   const onOpenDrawer = useCallback(() => {
     tracer.trace("Opening drawer", { hasFirmware: !!firmware, function: "onOpenDrawer" });
@@ -95,7 +97,9 @@ const FirmwareUpdate = (props: Props) => {
       firmwareName: firmware.final.name,
     });
 
-    setFirmwareUpdateOpened(true); // Prevents manager from reacting to device changes (?)
+    // Prevents manager from reacting to device changes
+    setPreventResetOnDeviceChange(true);
+
     const updateModalProps: UpdateModalProps = {
       withAppsToReinstall:
         !!installed &&
@@ -115,7 +119,6 @@ const FirmwareUpdate = (props: Props) => {
       device: device,
       error: error,
       deviceModelId: deviceSpecs.id,
-      setFirmwareUpdateOpened,
       setFirmwareUpdateCompleted,
       shouldReloadManagerOnCloseIfUpdateRefused: true,
     };
@@ -137,7 +140,7 @@ const FirmwareUpdate = (props: Props) => {
     onRequestClose,
     setDrawer,
     setFirmwareUpdateCompleted,
-    setFirmwareUpdateOpened,
+    setPreventResetOnDeviceChange,
     stepId,
     tracer,
   ]);
@@ -161,7 +164,7 @@ const FirmwareUpdate = (props: Props) => {
       <FirmwareUpdateBanner
         old
         right={
-          <FakeLink onClick={() => openURL(urls.contactSupport)}>
+          <FakeLink onClick={() => openURL(contactSupportUrl)}>
             <Trans i18nKey="manager.firmware.banner.old.cta" />
           </FakeLink>
         }