+++ apps/ledger-live-mobile/src/families/elrond/components/Flows/Delegate/components/SetDelegation/SetDelegation.tsx
@@ -234,10 +234,11 @@ const SetDelegation = (props: SetDelegationPropsType) => {
     <SafeAreaView style={[styles.root, { backgroundColor: colors.background }]}>
       <TrackScreen
         category="DelegationFlow"
-        name="Summary"
+        name={route.params.skipStartedStep ? "Step Starter" : "Summary"}
+        screen="Summary"
         flow="stake"
         action="delegate"
-        currency="egld"
+        currency="MultiversX"
       />
 
       <View style={styles.body}>