+++ libs/ledger-live-common/src/families/hedera/js-estimateMaxSpendable.ts
@@ -1,11 +1,11 @@
 import BigNumber from "bignumber.js";
-import { getEstimatedFees } from "./utils";
 import type { Account, AccountLike } from "@ledgerhq/types-live";
+import { getMainAccount } from "../../account/index";
 import type { Transaction } from "./types";
+import { getEstimatedFees } from "./utils";
 
 export default async function estimateMaxSpendable({
   account,
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   parentAccount,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   transaction,
@@ -16,7 +16,8 @@ export default async function estimateMaxSpendable({
 }): Promise<BigNumber> {
   const balance = account.balance;
 
-  const estimatedFees = await getEstimatedFees();
+  const mainAccount = getMainAccount(account, parentAccount);
+  const estimatedFees = await getEstimatedFees(mainAccount);
 
   let maxSpendable = balance.minus(estimatedFees);