+++ libs/ledger-live-common/src/families/solana/banner.test.ts
@@ -102,6 +102,7 @@ const account: SolanaAccount = {
         withdrawable: 0,
       },
     ],
+    unstakeReserve: BigNumber(0),
   },
 };